requirements.txt
google-search-results
langchain
langchain-openai
langchain-community
python-dotenv